- Se modificó el lexer para que pueda soportar numeros hexadecimales, binarios, Octadecimales, Exponenciales. - Se modificó el lexer para que acepte _ entre los numeros. - Se creó un pool de registros para poder usar registros temporales, por ejemplo al hacer una suma los registros que sean temporales en esa suma podran ser reutilizados para usar la menor cantidad de Registros posibles. - Se hizo limpieza en el código para borrar lineas de códigos duplicadas o con codigo viejo comentariado, para mantenerlo más limpio. - Se mejoró compatibilidad cambiando int por size_t, uint32_t, uint64_t, etc. - Se mejoró la forma de funcionar la vm, para poder soportar stack trace para las futuras llamadas a funciones, los futuros try-catch, y la futura librería debug. - También se mejoraron los errores del runtime para no imprimirse inmediatamente, si no guardar la info en la vm y al final imprimirlo.

+++ include/m.h
@@ -25,6 +25,6 @@ typedef enum /*ValueType*/
 #define MIN_STACK 20
 
 #define cast(t, expr)   ((t) (expr))
-#define cast_int(expr)  cast(int, expr)
+#define cast_int(expr)  cast(int64_t, expr)
 #define cast_double(expr)   cast(double, expr)
